pyproject.toml
# pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "BASF Backend Engineering Case Study"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.2"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.21"}
pydantic = {extras = ["email"], version = "^2.4.2"}
pydantic-settings = "^2.0.3"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "^4.0.1"  # Explicit bcrypt dependency with correct version
python-multipart = "^0.0.6"
python-dotenv = "^1.0.0"
asyncpg = "^0.28.0"  # PostgreSQL driver
aiosqlite = "^0.19.0"  # SQLite driver
alembic = "^1.12.0"
tenacity = "^8.2.3"
loguru = "^0.7.2"
celery = "^5.3.4"
redis = "^5.0.1"
httpx = "^0.25.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pytest-asyncio = "^0.21.1"
httpx = "^0.25.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"